% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{function1}
\alias{function1}
\title{A function}
\usage{
function1(x)
}
\arguments{
\item{x}{(numeric) input number}
}
\description{
This function adds one to the input.
}
\examples{
function1(2) #becomes 3
}
\keyword{add}
