% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{function2}
\alias{function2}
\title{Another function}
\usage{
function2(x)
}
\arguments{
\item{x}{(numeric) input number}
}
\description{
This function adds two to the input
}
\examples{
function2(2) #becomes 4
}
\keyword{add}
